sync

+++ database/migrations/2024_05_09_072909_create_subjects_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('subjects', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('number_sessions_per_week');
+            $table->double('number_sessions_per_week');
             $table->foreignId('semester_id')->references('id')->on('semesters')->onDelete('cascade');
             $table->timestamps();
         });
