+++ database/migrations/2024_05_09_073212_create_marks_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('marks', function (Blueprint $table) {
             $table->id();
-            $table->integer('result');
+            $table->double('result');
             $table->date('date');
             $table->foreignId('subject_id')->constrained('subjects')->onDelete('cascade');
             $table->foreignId('student_id')->constrained('students')->onDelete('cascade');
